Compass + Settingss Shortcut
Added a compass that uses the locationManager didUpdateHeading delegate method to update the compass image. I should not always point north.

If the user declined authorization for location services we want an easy access shortcut to phone settings. When app loads or when authorisation status gets updated the settingsShortcutButton should show/hide accordingly.