Annotation + Refactor
Added an annotation using a hardcoded location coordinates. The annotation is presented as a bubble around the location, without a pin. 
The idea is that users can set a bubble radius and color for each reminder.

Refactored the Add/EditReminderController into one: ReminderController by adding a modeSelected enum that allows setting up the view and navigation bar items.