DataModel Update
Made changes to the model that reflects the information that will be used in CoreData and used for stubbing. 

Area trigger has been changed to true/false as it made no since triggering the same message for entering and leaving an area. BorderColor has been renamed to bubbleColour and will be stored as a String safely using Enums.

Next: Build ReminderController view with all selectable properties.