UITextField
The input possibilities for the UITextFields are limited in several ways. 

The number of characters allowed, for respectively the title and message input fields, are currently set to 20 and 40.

The input fields for latitude and longitude only allow typing characters that could represent a value for these inputs. If a user enters a value that is outside the range of a possible minimum or maximum value for either  latitude or longitude they are replaced by the nearest possible value.