ReminderController
Added some safety to this screen by setting most Strings which are used once or multiple times to have one reference namely the structs PlaceHolderText and ToggleText. This als allows for easy editing since all the editing can be done in one location: Constants.swift.

The UITextFields for locationInfo and bubbleRadiusInfo are set to .isUserEnabled = false to prevent tapping and edit.