checkCoordinateInput + saveReminder
The checkCoordinateInput method has been updated to check if there is a connection. If there is it will perform its check, if not it will present an alert.

The information describing a reminder can now be saved to CoreData. Needs testing. 

Next: Update ActiveReminderController to present saved reminders instead of stub.