ReminderController + ArrowImage Animation
ReminderController is now set up to handle editing. Edits to reminders should now be properly reflected in the activeRemindersController.

Added an animation to the ArrowImage to better illustrate when the notification will be triggered. (Might needs some tweaking to really shine)

Todo: Add some extra colours to the bubbleColour options.
Next: Update map so it displays all currently active reminders