LocationNameFinder Backup
All code used to search for location name using coordinates will be deleted. This is now handled the other way around. User enters location name in a UISearchBar with auto complete which generates a latitude and longitude.

This commit is a backup for that code in case it will be used in a future project