MKMapView Reset
Every time a reminder is being added, edited or deleted the MainController needs to be notified of that change and reload all map data. Currently the ReminderController and the ActiveRemindersController both send out a notification when either of these actions occur. The MainController is set as an observer and listens for this notification. When notified it should print a log statement.

Test if statement gets printed.

If it does we need to refresh all reminders:
- Clear all current annotations and associated objects
- Refresh reminders array
- Create new reminders