fetchReminder + locationSearchBar
Created a fetchReminder method that fetches the reminder that belongs to the monitored region causing the trigger. The information associated with the reminder is then used to populate the notification.

Fixed a small issue where locationSearchBar would not properly reset its text when a new reminder is being added. Instead of setting the locationSearchBar.placeholder text the locationSearchBar.text was set to its placeholder text. This issue resulted in the user first having to tap the clear button before being able to start typing instead of doing it automatically.