Refactor
- Cleaned unused code and obsolete comments. 
- Merged createGeoFence and createNotificationContent methods into createNotification do prevent repeating region code.

Issue: Currently, reminders that have "trigger once" selected get added again whenever a reminder gets edited. Adding a property "isActive" to reminder should fix this. After updating or editing a reminder before adding a reminder to be monitored this property gets checked. If isActive equals to false it won't get monitored. Also when a notification has triggered for a reminder that has "repeat once" selected this property should be set to false.