isActive property  + Reminder Alert Message
Added a property to Reminder so user is able to activate and deactivate reminders. This allows user to manage clutter on map and keep reminders in list for later use. When notification trigger, and if the reminders isRepeat == false, the isActive property is set to false automatically. This fixes the issue of re-adding one use reminders to be monitored again when adding or editing a reminder.

The message displayed in the alert when the app is active now matches the one in the notification when the app is not in foreground.